Componentised some of the homepage
