+++ src/lib/Navbar.svelte
@@ -1,0 +1,32 @@
+<nav>
+    <span class="shadow">
+        <a href="/">Home</a>
+    </span>
+    <span class="shadow">
+        <a href="/about">About</a>
+    </span>
+
+    <span class="shadow">
+        <a href="/sponsors">DDDSW 2023</a>
+    </span>
+</nav>
+
+<style>
+    	nav {
+		display: flex;
+		justify-content: flex-end;
+		gap: 50px;
+		margin: 50px;
+	}
+
+	nav a {
+		color: white;
+		text-decoration: none;
+		font-size: x-large;
+	}
+
+    .shadow {
+		box-shadow: 0px 0px 50px 43px rgb(0 0 0 / 40%);
+		background-color: rgb(0 0 0 / 40%);
+	}
+</style>
